pyproject.toml
[project]
name = "al-edtech-backend-web-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi[all]==0.103.2",
    "pydantic==2.11.5",
    # Env & Tooling
    "python-dotenv==1.1.0",
    "poetry==1.8.2",
    "pylint==3.2.0",
    "toml==0.10.2",
    # AWS + Email
    "boto3==1.34.89",
    "sib-api-v3-sdk==7.6.0",
    # Server + Request handling
    "uvicorn[standard]==0.23.2",
    "python-multipart==0.0.9",
    # Security & Auth
    "bcrypt==4.1.3",
    "email-validator==2.1.1",
    "itsdangerous==2.2.0",
    "authlib==1.3.1",
    "PyJWT==2.8.0",
    # Database & ORM
    "sqlalchemy==2.0.29",
    "databases==0.9.0",
    "psycopg2-binary==2.9.9",
    "asyncpg==0.29.0",
    "pymongo==4.7.2",
    # Messaging/Notifications
    "pyfcm==2.0.4",
    "firebase_admin==6.5.0",
    # Utils
    "nanoid==2.0.0",
    "ragmaster",
    "tavily-python>=0.7.9",
    "langgraph>=0.5.2",
    "langchain-community>=0.3.21",
    "langsmith>=0.3.45",
    "alembic>=1.16.5",
    "motor>=3.5.3",
    "aioboto3>=13.1.1",
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "extractipy",
]
 
[tool.uv.sources]
ragmaster = { git = "https://github.com/Acelucid-Technologies/ragmaster.git" }
extractipy = { git = "https://github.com/Acelucid-Technologies/extractipy.git", rev = "v2.2.0" }